chore: refactor action buttons to it's own component
