+++ packages/chat-component/src/index.ts
@@ -6,5 +6,6 @@ export * from './components/document-previewer.js';
 export * from './components/loading-indicator.js';
 export * from './components/citation-list.js';
 export * from './components/chat-thread-component.js';
+export * from './components/chat-action-button.js';
 export * from './core/index.js';
 export * from './utils/index.js';
